refactor: quiz-note 페이지 리팩토링 및 공통 컴포넌트 분리
