+++ src/views/quiz-note/components/header.tsx
@@ -1,16 +1,17 @@
 'use client'
 
-import Icon, { IconProps } from '@/shared/components/icon'
-import IconButton from './icon-button'
+import Icon from '@/shared/components/icon'
 import { useQuizNoteContext } from '../context/quiz-note-context'
 import Text from '@/shared/components/ui/text'
 import { useEffect, useState } from 'react'
 import { cn } from '@/shared/lib/utils'
 import FolderSelectDrawer from './folder-select-drawer'
+import SortIconBtn from './sort-icon-button'
+import Link from 'next/link'
+import MenuDotsBtn from './menu-dots-button'
 
 // Header 컴포넌트
 const Header = () => {
-  const iconList = ['search', 'sort', 'menu-dots'] as IconProps['name'][]
   const { setSelectedFolderId, isSelectMode, setIsSelectMode } = useQuizNoteContext()
   const [isDrawerOpen, setIsDrawerOpen] = useState(false)
 
@@ -43,9 +44,13 @@ const Header = () => {
 
               {!isDrawerOpen && (
                 <div className="flex size-fit items-center gap-[16px]">
-                  {iconList.map((iconName) => (
-                    <IconButton key={iconName} iconName={iconName} />
-                  ))}
+                  <Link href="/note/search">
+                    <Icon name="search" className="size-[24px]" />
+                  </Link>
+
+                  <SortIconBtn />
+
+                  <MenuDotsBtn />
                 </div>
               )}
             </>
